requirements.txt
google-genai>=1.0.0
markdown>=3.5
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
